Add release plugin configuration
We need to enable all of the OS-specific profiles so the version numbers
of all the modules are bumped.

In turn this means we need to avoid calling the verify target during
release:prepare, as we can't build all the modules on any platform.

Similarly, we can't actually call the deploy target during
release:perform. We instead call `mvn -Prelease deploy` on each
supported operating system/architecture manually.