Improve code for populating openrs2.jawtDirectory
The previous profile-based logic didn't work on Java 8 due to the
architecture-specific lib directories. It was also complicated by the
fact that nar.javaHome isn't set at profile activation time unless the
user manually specifies it with -D on the Maven command line.

This commit replaces it with an Ant task that runs during the
initialization phase. It finds the first (and probably only) jawt
library with glob matching, allowing us to search architecture-specific
lib directories.

If found, it strips the file name to get just the directory name and
sets openrs2.jawtProperty. If not found, it makes the build fail.